requirements.txt
baostock==0.8.8
pandas==2.0.3
numpy==1.24.3
ta-lib==0.4.28
requests==2.31.0
beautifulsoup4==4.12.2
langchain==0.0.350
transformers==4.36.2
torch==2.1.2
python-dotenv==1.0.0 